selectors are now only exported from the selectors module
